Split utilities up into a constants and config folder
