+++ src/components/CreateContract/container.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { createContract } from '../../services/contractService';
 import swal from 'sweetalert2';
-import { Toast } from '../../utilities/Notifications';
+import { Toast } from '../../config/Notifications';
 
 export default Wrapped =>
   class extends Component {
