+++ src/components/NewBookingModal/UpdateBooking/container.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { PropTypes as PT } from 'prop-types';
 import { updateEvent, cancelEvent } from '../../../services/eventService';
 import { getHolidayStats } from '../../../services/holidayService';
-import { Toast } from '../../../utilities/Notifications';
+import { Toast } from '../../../config/Notifications';
 import Swal from 'sweetalert2';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
