+++ src/components/UserDetails/index.jsx
@@ -3,7 +3,7 @@ import { PropTypes as PT } from 'prop-types';
 import container from './container';
 import { Card } from '../common';
 import { StyleContainer } from './styled';
-import employeeRoles from '../../utilities/roles';
+import employeeRoles from '../../constants/roles';
 import FontAwesomeIcon from '@fortawesome/react-fontawesome';
 import { faKey } from '@fortawesome/fontawesome-free-solid';
