+++ src/constants/eventsView.js
