+++ src/services/eventService.js
@@ -1,4 +1,4 @@
-import axios from '../utilities/AxiosInstance';
+import axios from '../config/AxiosInstance';
 
 export const getAllEvents = () => {
   return axios.get('/Event/');
