+++ src/services/holidayService.js
@@ -1,4 +1,4 @@
-import axios from '../utilities/AxiosInstance';
+import axios from '../config/AxiosInstance';
 
 export const getHolidayStats = employeeId => {
   return axios.get(`/Event/findEmployeeHolidayStats/${employeeId}`);
